# maintence-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Maintence API",
        "description": "API Documentation for Maintenance System",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parts": {
            "get": {
                "tags": [
                    "Parts"
                ],
                "summary": "Get All Parts (Chunked)",
                "description": "Retrieve parts in chunks of 10. Use 'page' parameter to navigate.",
                "operationId": "c5500cfce11719821326175e7a7e415f",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Chunk number (default 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of parts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PartResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Parts"
                ],
                "summary": "Create New Part",
                "operationId": "c3f25c1eca998f26f4b14c692236831e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price_default",
                                    "amount",
                                    "warning_amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Compressor Valve"
                                    },
                                    "price_default": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "warning_amount": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Part created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Created"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/parts/{id}": {
            "get": {
                "tags": [
                    "Parts"
                ],
                "summary": "Get Part Details",
                "operationId": "46b9b619ef69fbfb66ca0c503b5af5d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Parts"
                ],
                "summary": "Update Part",
                "operationId": "9e8a91c1c454352e3ebdd357f0a93fbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Compressor Valve Updated"
                                    },
                                    "price_default": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 160
                                    },
                                    "amount": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "warning_amount": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Part updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Updated"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PartResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Parts"
                ],
                "summary": "Delete Part",
                "operationId": "369cb1877090cf19e309bbbc84773727",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Part deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Part deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PartResource": {
                "title": "Part Resource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "price_default": {
                        "type": "number",
                        "format": "float"
                    },
                    "amount": {
                        "type": "integer"
                    },
                    "warning_amount": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}